requirements.txt
Flask==2.3.2
Pyrebase4==4.7.0